+++ libs/render_delegate/render_delegate.cpp
@@ -479,9 +479,11 @@ HdArnoldRenderDelegate::HdArnoldRenderDelegate(bool isBatch, const TfToken &cont
         AiADPAddProductMetadata(AI_ADP_PLUGINVERSION, AtString{AI_VERSION});
         AiADPAddProductMetadata(AI_ADP_HOSTNAME, AtString{"Hydra"});
         AiADPAddProductMetadata(AI_ADP_HOSTVERSION, AtString{PXR_VERSION_STR});
-        // TODO(pal): We need to investigate if it's safe to set session to AI_SESSION_BATCH when rendering in husk for
-        //  example. ie. is husk creating a separate render delegate for each frame, or syncs the changes?
-        AiBegin(AI_SESSION_INTERACTIVE);
+        if (context == str::t_husk) {
+            AiBegin(AI_SESSION_BATCH);
+        } else {
+            AiBegin(AI_SESSION_INTERACTIVE);
+        }
     }
     _supportedRprimTypes = {HdPrimTypeTokens->mesh, HdPrimTypeTokens->volume, HdPrimTypeTokens->points,
                             HdPrimTypeTokens->basisCurves, str::t_procedural_custom};
@@ -549,7 +551,7 @@ HdArnoldRenderDelegate::HdArnoldRenderDelegate(bool isBatch, const TfToken &cont
 
     if (_renderDelegateOwnsUniverse) {
         _universe = AiUniverse();
-        _renderSession = AiRenderSession(_universe, AI_SESSION_INTERACTIVE);
+        _renderSession = AiRenderSession(_universe, (context == str::t_husk) ? AI_SESSION_BATCH : AI_SESSION_INTERACTIVE);
     }
 
     _renderParam.reset(new HdArnoldRenderParam(this));